(important) from int64 to float64

+++ builder/options.go
@@ -1,21 +1,21 @@
 // SPDX-License-Identifier: MIT
 // Package: lvlath/builder
 //
-// options.go — functional options for the builder package.
+// options.go - functional options for the builder package.
 //
 // Contract (strict):
-//   • Options are functional (type Option func(*builderConfig)).
-//   • Option constructors VALIDATE and PANIC on meaningless inputs
+//   - Options are functional (type Option func(*builderConfig)).
+//   - Option constructors VALIDATE and PANIC on meaningless inputs
 //     (per lvlath 99-rules). Algorithms themselves MUST NOT panic.
-//   • Determinism is explicit: seeding is done via WithSeed or WithRand.
-//   • No hidden globals; everything flows through builderConfig.
+//   - Determinism is explicit: seeding is done via WithSeed or WithRand.
+//   - No hidden globals; everything flows through builderConfig.
 //
 // AI-Hints:
-//   • Prefer WithSeed for reproducible stochastic builders (Random*).
-//   • Use WithIDScheme to align vertex labels across tests/golden files.
-//   • WithPartitionPrefix controls K_{m,n} labels; empty values mean
+//   - Prefer WithSeed for reproducible stochastic builders (Random*).
+//   - Use WithIDScheme to align vertex labels across tests/golden files.
+//   - WithPartitionPrefix controls K_{m,n} labels; empty values mean
 //     “use defaults”, not an error (deterministic fallback).
-//   • WithWeightFn affects weighted graphs only; core controls whether
+//   - WithWeightFn affects weighted graphs only; core controls whether
 //     weights are observed.
 
 package builder
@@ -37,6 +37,7 @@ func WithIDScheme(fn func(int) string) BuilderOption {
 		// Fail fast: option constructors validate and panic (99-rules).
 		panic("builder: WithIDScheme(nil)")
 	}
+
 	return func(c *builderConfig) {
 		// Assign the provided function; used by all topology builders.
 		c.idFn = fn
@@ -71,7 +72,7 @@ func WithSeed(seed int64) BuilderOption {
 // The function receives the (possibly nil) RNG and MUST be pure w.r.t.
 // input RNG state to preserve determinism. Panics on nil.
 // Complexity: O(1) time, O(1) space.
-func WithWeightFn(fn func(*rand.Rand) int64) BuilderOption {
+func WithWeightFn(fn func(*rand.Rand) float64) BuilderOption {
 	if fn == nil {
 		// Fail fast; weight policy must be explicit if customized.
 		panic("builder: WithWeightFn(nil)")
@@ -99,6 +100,7 @@ func WithAmplitude(A float64) BuilderOption {
 	if A <= 0 {
 		panic("builder: WithAmplitude(A<=0)")
 	}
+
 	return func(c *builderConfig) {
 		// Deterministic scalar controlling signal scale.
 		c.amplitude = A
@@ -112,6 +114,7 @@ func WithFrequency(f0 float64) BuilderOption {
 	if f0 <= 0 {
 		panic("builder: WithFrequency(f0<=0)")
 	}
+
 	return func(c *builderConfig) {
 		// Fundamental frequency parameter for signal synthesis.
 		c.frequency = f0
@@ -135,6 +138,7 @@ func WithNoise(sigma float64) BuilderOption {
 	if sigma < 0 {
 		panic("builder: WithNoise(sigma<0)")
 	}
+
 	return func(c *builderConfig) {
 		// Standard deviation for additive noise; 0 means noiseless.
 		c.noiseSigma = sigma
